Parallel_Backend_Bot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9

redis==5.0.4

passlib[bcrypt]==1.7.4
bcrypt==4.0.1          
python-jose==3.3.0     
email-validator==2.1.1

python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

ib_async==2.0.1
